DAQ-1442 Avoid blocking the update method performing the update work
The update method now submits a job to a single threaded single queue
element Executor. The work needed to perform the update is done by the
executor before submitting to another queue for plotting.

If another update request is received a maximum of one will be queued to
ensure the GUI reflects the latest state.

Change-Id: I3c96c2847d175f6828cc6c765ba975bc887a4299